requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
contourpy==1.0.6
cycler==0.11.0
Django==3.2.11
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Flask==2.1.1
fonttools==4.38.0
frozenlist==1.3.0
future==0.18.2
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.4
gspread==5.3.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.4
line-bot-sdk==2.2.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mne==0.23.4
multidict==6.0.2
numpy==1.21.2
oauth2client==4.1.3
oauthlib==3.2.0
openai==0.27.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
PyJWT==1.7.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
yarl==1.7.2
zipp==3.8.0
